management/server/requirements.txt
Flask==3.1.0
flask_cors==5.0.1
mysql-connector-python==9.2.0
pycryptodomex==3.20.0
tabulate==0.9.0
Werkzeug==3.1.3
PyJWT==2.10.1
dotenv==0.9.9
magic-pdf==1.3.12
transformers==4.49.0
elasticsearch==8.12.0
minio==7.2.4
strenum==0.4.15
peewee==3.17.1
pytz==2020.5
bs4==0.0.2
opencv-python==4.10.0.84
ultralytics==8.3.102
doclayout-yolo==0.0.3
ftfy==6.3.1
dill==0.4.0
shapely==2.1.0
pyclipper==1.3.0.post6
omegaconf==2.3.0
rapid-table==1.0.3
openai==1.70.0
redis==6.2.0